+++ src/apsg/database/_alchemy.py
@@ -22,7 +22,6 @@ from sqlalchemy import (
     Table,
     Text,
     text,
-    UniqueConstraint,
 )
 from sqlalchemy.orm import relationship
 from sqlalchemy.ext.declarative import declarative_base